fix: increase the number of significant digits for VAF
The issue was that whenever a VAF under 0.001 was reported, it would be
rounded to 0. Increasing the number of significant digits addresses this
issue, but it also leads to somewhat of a misrepresentation of the
precision of the analysis. For now though, it is more important that
these variants of low frequency are represented accurately. Whether or
not we want to do a conditional formatting of VAF values, similar to
what we had in v0.1.3, will have to be a question we address in the
future.