feat: show points for all samples
Previously, if the VAF for a sample was below the limit of detection a
point would not be drawn for this sample in the plot. This was not a
problem since the x-axis labels made it apparent that there was sampling
done at this time. With the cleanup of the labels in the previous
release this is no longer the case.

This PR makes it so that points are plotted for all samples, regardless
of whether they are below the limit of detection or not.