fix: use HGVSc in plots if HGVSp is unknown
For the panel titles in the plot the HGVSc annotation has been used.
However, in the case of splice variants, the protein change is unknown
and represented as "p.?". Since this is not very informative, we instead
opt to show the DNA modification in cases where HGVSp matches the string
"p.?". The recommendation from HGVS is to use this nomenclature for
splice variants, and we just have to trust that the annotations follow
this convention.