fix: check that annotation dates are defined (#44)
If for some reason the date for an annotation is missing, or couldn't be
parsed, the date will be obtained from the previously defined date in
the sheet. This adds a check to make sure that dates for all annotations
are defined before filling in missing dates for other rows.